Require all guide baseline test data to match closely.
