+++ tests/components/optics/test_guide_baseline.py
@@ -9,8 +9,6 @@ from mcni import neutron_buffer, neutron
 from mcni.neutron_storage import neutrons_as_npyarr, ndblsperneutron
 from mcvine import run_script
 from mcvine.acc import test
-from mcvine.acc.components.optics.guide import Guide
-from mcvine.acc.geometry.plane import Plane
 
 
 thisdir = os.path.dirname(__file__)
@@ -61,43 +59,11 @@ def test_compare_mcvine(num_neutrons = int(1e6), debug=False):
         plotHist(Ixy)
         plotHist(Ixdivx)
         plotHist((Ixy-mcvine_Ixy)/mcvine_Ixy)
-    assert histogram_is_close(
-        Ixy.I, mcvine_Ixy.I,
-        min_nonzero_fraction=0.92,
-        rtol=1e-3, min_rdiff_fraction=0.98,
-        atol=1e-7*1e-3, min_adiff_fraction=0.99
-    )
-    assert histogram_is_close(
-        Ixdivx.I, mcvine_Ixdivx.I,
-        min_nonzero_fraction=0.2,
-        rtol=1e-3, min_rdiff_fraction=0.99,
-        atol=1e-7*1e-3, min_adiff_fraction=0.99
-    )
-    return
+    assert mcvine_Ixy.shape() == Ixy.shape()
+    assert mcvine_Ixdivx.shape() == Ixdivx.shape()
+    assert np.allclose(mcvine_Ixy.data().storage(), Ixy.data().storage())
+    assert np.allclose(mcvine_Ixdivx.data().storage(), Ixdivx.data().storage())
 
-def histogram_is_close(
-        h1, h2, min_nonzero_fraction=0.9,
-        rtol=0.001, min_rdiff_fraction=0.9,
-        atol=1e-10, min_adiff_fraction=0.9,
-):
-    "check if histogram h1 is close to the reference h2"
-    if h2.shape != h1.shape: return False
-    rdiff = h1/h2-1
-    isfinite = np.isfinite(rdiff)
-    rdiff = rdiff[isfinite]
-    adiff = (h1-h2)[~isfinite]
-    # make sure there is enough data points that are not zero
-    print(rdiff.size, h2.size)
-    if rdiff.size<min_nonzero_fraction*h2.size: return False
-    # make sure for a good portion of nonzero data points,
-    # the relative diff is smaller than the given tolerance
-    print ((np.abs(rdiff)<rtol).sum() , rdiff.size)
-    if (np.abs(rdiff)<rtol).sum() < rdiff.size*min_rdiff_fraction:
-        return False
-    print ((np.abs(adiff)<atol).sum() , adiff.size)
-    if (np.abs(adiff)<atol).sum() < adiff.size*min_adiff_fraction:
-        return False
-    return True
 
 def debug():
     global interactive
